Renamed reconfiguration Filenames
