+++ src/modules/reconfiguration/CMakeLists.txt
@@ -16,5 +16,5 @@ forte_add_module(RECONFIGURATION OFF "FORTE Online Reconfiguration")
 #############################################################################
 # FORTE Reconfiguration FBs
 #############################################################################
-forte_add_sourcefile_hcpp(EC_KILL_ELEM EC_SET_EVT EC_START_ELEM EC_STOP_ELEM)
-forte_add_sourcefile_hcpp(ST_CREATE_CONN ST_CREATE_FB ST_DEL_CONN ST_DEL_FB ST_REC_CONN ST_SET_PARM)
+forte_add_sourcefile_hcpp(EC_KILL_ELEM_fbt EC_SET_EVT_fbt EC_START_ELEM_fbt EC_STOP_ELEM_fbt)
+forte_add_sourcefile_hcpp(ST_CREATE_CONN_fbt ST_CREATE_FB_fbt ST_DEL_CONN_fbt ST_DEL_FB_fbt ST_REC_CONN_fbt ST_SET_PARM_fbt)
