+++ src/modules/reconfiguration/EC_KILL_ELEM_fbt.cpp
@@ -9,30 +9,30 @@
  * Contributors:
  *   Matthias Plasch, Gerhard Ebenhofer, Alois Zoitl - initial API and implementation and/or initial documentation
  *******************************************************************************/
-#include "EC_START_ELEM.h"
+#include "EC_KILL_ELEM_fbt.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "EC_START_ELEM_gen.cpp"
+#include "EC_KILL_ELEM_fbt_gen.cpp"
 #endif
 
 #include "criticalregion.h"
 #include "resource.h"
 
-#include "device.h"
 #include "../../stdfblib/ita/DEV_MGR.h"
+#include "device.h"
 
-DEFINE_FIRMWARE_FB(FORTE_EC_START_ELEM, g_nStringIdEC_START_ELEM)
+DEFINE_FIRMWARE_FB(FORTE_EC_KILL_ELEM, g_nStringIdEC_KILL_ELEM)
 
-const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataInputNames[] = {g_nStringIdQI, g_nStringIdELEM_NAME, g_nStringIdDST};
-const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataInputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING, g_nStringIdWSTRING};
-const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataOutputNames[] = {g_nStringIdQO, g_nStringIdSTATUS};
-const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataOutputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING};
-const TDataIOID FORTE_EC_START_ELEM::scmEIWith[] = {1, 2, 0, scmWithListDelimiter};
-const TForteInt16 FORTE_EC_START_ELEM::scmEIWithIndexes[] = {0};
-const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmEventInputNames[] = {g_nStringIdREQ};
-const TDataIOID FORTE_EC_START_ELEM::scmEOWith[] = {1, 0, scmWithListDelimiter};
-const TForteInt16 FORTE_EC_START_ELEM::scmEOWithIndexes[] = {0};
-const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmEventOutputNames[] = {g_nStringIdCNF};
-const SFBInterfaceSpec FORTE_EC_START_ELEM::scmFBInterfaceSpec = {
+const CStringDictionary::TStringId FORTE_EC_KILL_ELEM::scmDataInputNames[] = {g_nStringIdQI, g_nStringIdELEM_NAME, g_nStringIdDST};
+const CStringDictionary::TStringId FORTE_EC_KILL_ELEM::scmDataInputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING, g_nStringIdWSTRING};
+const CStringDictionary::TStringId FORTE_EC_KILL_ELEM::scmDataOutputNames[] = {g_nStringIdQO, g_nStringIdSTATUS};
+const CStringDictionary::TStringId FORTE_EC_KILL_ELEM::scmDataOutputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING};
+const TDataIOID FORTE_EC_KILL_ELEM::scmEIWith[] = {1, 2, 0, scmWithListDelimiter};
+const TForteInt16 FORTE_EC_KILL_ELEM::scmEIWithIndexes[] = {0};
+const CStringDictionary::TStringId FORTE_EC_KILL_ELEM::scmEventInputNames[] = {g_nStringIdREQ};
+const TDataIOID FORTE_EC_KILL_ELEM::scmEOWith[] = {1, 0, scmWithListDelimiter};
+const TForteInt16 FORTE_EC_KILL_ELEM::scmEOWithIndexes[] = {0};
+const CStringDictionary::TStringId FORTE_EC_KILL_ELEM::scmEventOutputNames[] = {g_nStringIdCNF};
+const SFBInterfaceSpec FORTE_EC_KILL_ELEM::scmFBInterfaceSpec = {
   1, scmEventInputNames, scmEIWith, scmEIWithIndexes,
   1, scmEventOutputNames, scmEOWith, scmEOWithIndexes,
   3, scmDataInputNames, scmDataInputTypeIds,
@@ -41,7 +41,7 @@ const SFBInterfaceSpec FORTE_EC_START_ELEM::scmFBInterfaceSpec = {
   0, nullptr
 };
 
-FORTE_EC_START_ELEM::FORTE_EC_START_ELEM(const CStringDictionary::TStringId paInstanceNameId, forte::core::CFBContainer &paContainer) :
+FORTE_EC_KILL_ELEM::FORTE_EC_KILL_ELEM(const CStringDictionary::TStringId paInstanceNameId, forte::core::CFBContainer &paContainer) :
     CFunctionBlock(paContainer, &scmFBInterfaceSpec, paInstanceNameId),
     var_conn_QO(var_QO),
     var_conn_STATUS(var_STATUS),
@@ -53,7 +53,7 @@ FORTE_EC_START_ELEM::FORTE_EC_START_ELEM(const CStringDictionary::TStringId paIn
     conn_STATUS(this, 1, &var_conn_STATUS) {
 };
 
-void FORTE_EC_START_ELEM::setInitialValues() {
+void FORTE_EC_KILL_ELEM::setInitialValues() {
   var_QI = 0_BOOL;
   var_ELEM_NAME = u""_WSTRING;
   var_DST = u""_WSTRING;
@@ -61,7 +61,7 @@ void FORTE_EC_START_ELEM::setInitialValues() {
   var_STATUS = u""_WSTRING;
 }
 
-void FORTE_EC_START_ELEM::executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) {
+void FORTE_EC_KILL_ELEM::executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) {
   switch(paEIID) {
     case scmEventREQID:
       var_QO = var_QI;
@@ -75,12 +75,12 @@ void FORTE_EC_START_ELEM::executeEvent(TEventID paEIID, CEventChainExecutionThre
   }
 }
 
-void FORTE_EC_START_ELEM::executeRQST(){
+void FORTE_EC_KILL_ELEM::executeRQST(){
   forte::core::SManagementCMD theCommand;
 
   theCommand.mDestination = CStringDictionary::getInstance().getId(var_DST.getValue());
   theCommand.mFirstParam.pushBack(CStringDictionary::getInstance().getId(var_ELEM_NAME.getValue()));
-  theCommand.mCMD = EMGMCommandType::Start;
+  theCommand.mCMD = EMGMCommandType::Kill;
 
   EMGMResponse resp = getDevice()->executeMGMCommand(theCommand);
 
@@ -91,7 +91,7 @@ void FORTE_EC_START_ELEM::executeRQST(){
   var_STATUS = CIEC_WSTRING(retVal.c_str());
 }
 
-void FORTE_EC_START_ELEM::readInputData(TEventID paEIID) {
+void FORTE_EC_KILL_ELEM::readInputData(TEventID paEIID) {
   switch(paEIID) {
     case scmEventREQID: {
       readData(1, var_ELEM_NAME, conn_ELEM_NAME);
@@ -104,7 +104,7 @@ void FORTE_EC_START_ELEM::readInputData(TEventID paEIID) {
   }
 }
 
-void FORTE_EC_START_ELEM::writeOutputData(TEventID paEIID) {
+void FORTE_EC_KILL_ELEM::writeOutputData(TEventID paEIID) {
   switch(paEIID) {
     case scmEventCNFID: {
       writeData(1, var_STATUS, conn_STATUS);
@@ -116,7 +116,7 @@ void FORTE_EC_START_ELEM::writeOutputData(TEventID paEIID) {
   }
 }
 
-CIEC_ANY *FORTE_EC_START_ELEM::getDI(size_t paIndex) {
+CIEC_ANY *FORTE_EC_KILL_ELEM::getDI(size_t paIndex) {
   switch(paIndex) {
     case 0: return &var_QI;
     case 1: return &var_ELEM_NAME;
@@ -125,7 +125,7 @@ CIEC_ANY *FORTE_EC_START_ELEM::getDI(size_t paIndex) {
   return nullptr;
 }
 
-CIEC_ANY *FORTE_EC_START_ELEM::getDO(size_t paIndex) {
+CIEC_ANY *FORTE_EC_KILL_ELEM::getDO(size_t paIndex) {
   switch(paIndex) {
     case 0: return &var_QO;
     case 1: return &var_STATUS;
@@ -133,14 +133,15 @@ CIEC_ANY *FORTE_EC_START_ELEM::getDO(size_t paIndex) {
   return nullptr;
 }
 
-CEventConnection *FORTE_EC_START_ELEM::getEOConUnchecked(TPortId paIndex) {
+
+CEventConnection *FORTE_EC_KILL_ELEM::getEOConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_CNF;
   }
   return nullptr;
 }
 
-CDataConnection **FORTE_EC_START_ELEM::getDIConUnchecked(TPortId paIndex) {
+CDataConnection **FORTE_EC_KILL_ELEM::getDIConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_QI;
     case 1: return &conn_ELEM_NAME;
@@ -149,12 +150,11 @@ CDataConnection **FORTE_EC_START_ELEM::getDIConUnchecked(TPortId paIndex) {
   return nullptr;
 }
 
-CDataConnection *FORTE_EC_START_ELEM::getDOConUnchecked(TPortId paIndex) {
+CDataConnection *FORTE_EC_KILL_ELEM::getDOConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_QO;
     case 1: return &conn_STATUS;
   }
   return nullptr;
 }
-
