+++ src/modules/reconfiguration/EC_START_ELEM_fbt.cpp
@@ -9,9 +9,9 @@
  * Contributors:
  *   Matthias Plasch, Gerhard Ebenhofer, Alois Zoitl - initial API and implementation and/or initial documentation
  *******************************************************************************/
-#include "ST_SET_PARM.h"
+#include "EC_START_ELEM_fbt.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "ST_SET_PARM_gen.cpp"
+#include "EC_START_ELEM_fbt_gen.cpp"
 #endif
 
 #include "criticalregion.h"
@@ -20,52 +20,48 @@
 #include "device.h"
 #include "../../stdfblib/ita/DEV_MGR.h"
 
-DEFINE_FIRMWARE_FB(FORTE_ST_SET_PARM, g_nStringIdST_SET_PARM)
+DEFINE_FIRMWARE_FB(FORTE_EC_START_ELEM, g_nStringIdEC_START_ELEM)
 
-const CStringDictionary::TStringId FORTE_ST_SET_PARM::scmDataInputNames[] = {g_nStringIdQI, g_nStringIdELEM_NAME, g_nStringIdELEM_DATA_IN, g_nStringIdPARM_VAL, g_nStringIdDST};
-const CStringDictionary::TStringId FORTE_ST_SET_PARM::scmDataInputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING, g_nStringIdWSTRING, g_nStringIdWSTRING, g_nStringIdWSTRING};
-const CStringDictionary::TStringId FORTE_ST_SET_PARM::scmDataOutputNames[] = {g_nStringIdQO, g_nStringIdSTATUS};
-const CStringDictionary::TStringId FORTE_ST_SET_PARM::scmDataOutputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING};
-const TDataIOID FORTE_ST_SET_PARM::scmEIWith[] = {1, 2, 3, 4, 0, scmWithListDelimiter};
-const TForteInt16 FORTE_ST_SET_PARM::scmEIWithIndexes[] = {0};
-const CStringDictionary::TStringId FORTE_ST_SET_PARM::scmEventInputNames[] = {g_nStringIdREQ};
-const TDataIOID FORTE_ST_SET_PARM::scmEOWith[] = {1, 0, scmWithListDelimiter};
-const TForteInt16 FORTE_ST_SET_PARM::scmEOWithIndexes[] = {0};
-const CStringDictionary::TStringId FORTE_ST_SET_PARM::scmEventOutputNames[] = {g_nStringIdCNF};
-const SFBInterfaceSpec FORTE_ST_SET_PARM::scmFBInterfaceSpec = {
+const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataInputNames[] = {g_nStringIdQI, g_nStringIdELEM_NAME, g_nStringIdDST};
+const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataInputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING, g_nStringIdWSTRING};
+const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataOutputNames[] = {g_nStringIdQO, g_nStringIdSTATUS};
+const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmDataOutputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING};
+const TDataIOID FORTE_EC_START_ELEM::scmEIWith[] = {1, 2, 0, scmWithListDelimiter};
+const TForteInt16 FORTE_EC_START_ELEM::scmEIWithIndexes[] = {0};
+const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmEventInputNames[] = {g_nStringIdREQ};
+const TDataIOID FORTE_EC_START_ELEM::scmEOWith[] = {1, 0, scmWithListDelimiter};
+const TForteInt16 FORTE_EC_START_ELEM::scmEOWithIndexes[] = {0};
+const CStringDictionary::TStringId FORTE_EC_START_ELEM::scmEventOutputNames[] = {g_nStringIdCNF};
+const SFBInterfaceSpec FORTE_EC_START_ELEM::scmFBInterfaceSpec = {
   1, scmEventInputNames, scmEIWith, scmEIWithIndexes,
   1, scmEventOutputNames, scmEOWith, scmEOWithIndexes,
-  5, scmDataInputNames, scmDataInputTypeIds,
+  3, scmDataInputNames, scmDataInputTypeIds,
   2, scmDataOutputNames, scmDataOutputTypeIds,
   0, nullptr,
   0, nullptr
 };
 
-FORTE_ST_SET_PARM::FORTE_ST_SET_PARM(const CStringDictionary::TStringId paInstanceNameId, forte::core::CFBContainer &paContainer) :
+FORTE_EC_START_ELEM::FORTE_EC_START_ELEM(const CStringDictionary::TStringId paInstanceNameId, forte::core::CFBContainer &paContainer) :
     CFunctionBlock(paContainer, &scmFBInterfaceSpec, paInstanceNameId),
     var_conn_QO(var_QO),
     var_conn_STATUS(var_STATUS),
     conn_CNF(this, 0),
     conn_QI(nullptr),
     conn_ELEM_NAME(nullptr),
-    conn_ELEM_DATA_IN(nullptr),
-    conn_PARM_VAL(nullptr),
     conn_DST(nullptr),
     conn_QO(this, 0, &var_conn_QO),
     conn_STATUS(this, 1, &var_conn_STATUS) {
 };
 
-void FORTE_ST_SET_PARM::setInitialValues() {
+void FORTE_EC_START_ELEM::setInitialValues() {
   var_QI = 0_BOOL;
   var_ELEM_NAME = u""_WSTRING;
-  var_ELEM_DATA_IN = u""_WSTRING;
-  var_PARM_VAL = u""_WSTRING;
   var_DST = u""_WSTRING;
   var_QO = 0_BOOL;
   var_STATUS = u""_WSTRING;
 }
 
-void FORTE_ST_SET_PARM::executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) {
+void FORTE_EC_START_ELEM::executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) {
   switch(paEIID) {
     case scmEventREQID:
       var_QO = var_QI;
@@ -79,13 +75,12 @@ void FORTE_ST_SET_PARM::executeEvent(TEventID paEIID, CEventChainExecutionThread
   }
 }
 
-void FORTE_ST_SET_PARM::executeRQST() {
+void FORTE_EC_START_ELEM::executeRQST(){
   forte::core::SManagementCMD theCommand;
+
   theCommand.mDestination = CStringDictionary::getInstance().getId(var_DST.getValue());
   theCommand.mFirstParam.pushBack(CStringDictionary::getInstance().getId(var_ELEM_NAME.getValue()));
-  theCommand.mFirstParam.pushBack(CStringDictionary::getInstance().getId(var_ELEM_DATA_IN.getValue()));
-  theCommand.mAdditionalParams = func_WSTRING_TO_STRING(var_PARM_VAL);
-  theCommand.mCMD = EMGMCommandType::Write;
+  theCommand.mCMD = EMGMCommandType::Start;
 
   EMGMResponse resp = getDevice()->executeMGMCommand(theCommand);
 
@@ -96,13 +91,11 @@ void FORTE_ST_SET_PARM::executeRQST() {
   var_STATUS = CIEC_WSTRING(retVal.c_str());
 }
 
-void FORTE_ST_SET_PARM::readInputData(TEventID paEIID) {
+void FORTE_EC_START_ELEM::readInputData(TEventID paEIID) {
   switch(paEIID) {
     case scmEventREQID: {
       readData(1, var_ELEM_NAME, conn_ELEM_NAME);
-      readData(2, var_ELEM_DATA_IN, conn_ELEM_DATA_IN);
-      readData(3, var_PARM_VAL, conn_PARM_VAL);
-      readData(4, var_DST, conn_DST);
+      readData(2, var_DST, conn_DST);
       readData(0, var_QI, conn_QI);
       break;
     }
@@ -111,7 +104,7 @@ void FORTE_ST_SET_PARM::readInputData(TEventID paEIID) {
   }
 }
 
-void FORTE_ST_SET_PARM::writeOutputData(TEventID paEIID) {
+void FORTE_EC_START_ELEM::writeOutputData(TEventID paEIID) {
   switch(paEIID) {
     case scmEventCNFID: {
       writeData(1, var_STATUS, conn_STATUS);
@@ -123,18 +116,16 @@ void FORTE_ST_SET_PARM::writeOutputData(TEventID paEIID) {
   }
 }
 
-CIEC_ANY *FORTE_ST_SET_PARM::getDI(size_t paIndex) {
+CIEC_ANY *FORTE_EC_START_ELEM::getDI(size_t paIndex) {
   switch(paIndex) {
     case 0: return &var_QI;
     case 1: return &var_ELEM_NAME;
-    case 2: return &var_ELEM_DATA_IN;
-    case 3: return &var_PARM_VAL;
-    case 4: return &var_DST;
+    case 2: return &var_DST;
   }
   return nullptr;
 }
 
-CIEC_ANY *FORTE_ST_SET_PARM::getDO(size_t paIndex) {
+CIEC_ANY *FORTE_EC_START_ELEM::getDO(size_t paIndex) {
   switch(paIndex) {
     case 0: return &var_QO;
     case 1: return &var_STATUS;
@@ -142,25 +133,23 @@ CIEC_ANY *FORTE_ST_SET_PARM::getDO(size_t paIndex) {
   return nullptr;
 }
 
-CEventConnection *FORTE_ST_SET_PARM::getEOConUnchecked(TPortId paIndex) {
+CEventConnection *FORTE_EC_START_ELEM::getEOConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_CNF;
   }
   return nullptr;
 }
 
-CDataConnection **FORTE_ST_SET_PARM::getDIConUnchecked(TPortId paIndex) {
+CDataConnection **FORTE_EC_START_ELEM::getDIConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_QI;
     case 1: return &conn_ELEM_NAME;
-    case 2: return &conn_ELEM_DATA_IN;
-    case 3: return &conn_PARM_VAL;
-    case 4: return &conn_DST;
+    case 2: return &conn_DST;
   }
   return nullptr;
 }
 
-CDataConnection *FORTE_ST_SET_PARM::getDOConUnchecked(TPortId paIndex) {
+CDataConnection *FORTE_EC_START_ELEM::getDOConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_QO;
     case 1: return &conn_STATUS;
