+++ src/modules/reconfiguration/ST_DEL_FB_fbt.cpp
@@ -9,9 +9,9 @@
  * Contributors:
  *   Matthias Plasch, Gerhard Ebenhofer, Alois Zoitl - initial API and implementation and/or initial documentation
  *******************************************************************************/
-#include "EC_STOP_ELEM.h"
+#include "ST_DEL_FB_fbt.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "EC_STOP_ELEM_gen.cpp"
+#include "ST_DEL_FB_fbt_gen.cpp"
 #endif
 
 #include "criticalregion.h"
@@ -20,19 +20,19 @@
 #include "device.h"
 #include "../../stdfblib/ita/DEV_MGR.h"
 
-DEFINE_FIRMWARE_FB(FORTE_EC_STOP_ELEM, g_nStringIdEC_STOP_ELEM)
+DEFINE_FIRMWARE_FB(FORTE_ST_DEL_FB, g_nStringIdST_DEL_FB)
 
-const CStringDictionary::TStringId FORTE_EC_STOP_ELEM::scmDataInputNames[] = {g_nStringIdQI, g_nStringIdELEM_NAME, g_nStringIdDST};
-const CStringDictionary::TStringId FORTE_EC_STOP_ELEM::scmDataInputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING, g_nStringIdWSTRING};
-const CStringDictionary::TStringId FORTE_EC_STOP_ELEM::scmDataOutputNames[] = {g_nStringIdQO, g_nStringIdSTATUS};
-const CStringDictionary::TStringId FORTE_EC_STOP_ELEM::scmDataOutputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING};
-const TDataIOID FORTE_EC_STOP_ELEM::scmEIWith[] = {1, 2, 0, scmWithListDelimiter};
-const TForteInt16 FORTE_EC_STOP_ELEM::scmEIWithIndexes[] = {0};
-const CStringDictionary::TStringId FORTE_EC_STOP_ELEM::scmEventInputNames[] = {g_nStringIdREQ};
-const TDataIOID FORTE_EC_STOP_ELEM::scmEOWith[] = {1, 0, scmWithListDelimiter};
-const TForteInt16 FORTE_EC_STOP_ELEM::scmEOWithIndexes[] = {0};
-const CStringDictionary::TStringId FORTE_EC_STOP_ELEM::scmEventOutputNames[] = {g_nStringIdCNF};
-const SFBInterfaceSpec FORTE_EC_STOP_ELEM::scmFBInterfaceSpec = {
+const CStringDictionary::TStringId FORTE_ST_DEL_FB::scmDataInputNames[] = {g_nStringIdQI, g_nStringIdFB_NAME, g_nStringIdDST};
+const CStringDictionary::TStringId FORTE_ST_DEL_FB::scmDataInputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING, g_nStringIdWSTRING};
+const CStringDictionary::TStringId FORTE_ST_DEL_FB::scmDataOutputNames[] = {g_nStringIdQO, g_nStringIdSTATUS};
+const CStringDictionary::TStringId FORTE_ST_DEL_FB::scmDataOutputTypeIds[] = {g_nStringIdBOOL, g_nStringIdWSTRING};
+const TDataIOID FORTE_ST_DEL_FB::scmEIWith[] = {1, 2, 0, scmWithListDelimiter};
+const TForteInt16 FORTE_ST_DEL_FB::scmEIWithIndexes[] = {0};
+const CStringDictionary::TStringId FORTE_ST_DEL_FB::scmEventInputNames[] = {g_nStringIdREQ};
+const TDataIOID FORTE_ST_DEL_FB::scmEOWith[] = {1, 0, scmWithListDelimiter};
+const TForteInt16 FORTE_ST_DEL_FB::scmEOWithIndexes[] = {0};
+const CStringDictionary::TStringId FORTE_ST_DEL_FB::scmEventOutputNames[] = {g_nStringIdCNF};
+const SFBInterfaceSpec FORTE_ST_DEL_FB::scmFBInterfaceSpec = {
   1, scmEventInputNames, scmEIWith, scmEIWithIndexes,
   1, scmEventOutputNames, scmEOWith, scmEOWithIndexes,
   3, scmDataInputNames, scmDataInputTypeIds,
@@ -41,27 +41,27 @@ const SFBInterfaceSpec FORTE_EC_STOP_ELEM::scmFBInterfaceSpec = {
   0, nullptr
 };
 
-FORTE_EC_STOP_ELEM::FORTE_EC_STOP_ELEM(const CStringDictionary::TStringId paInstanceNameId, forte::core::CFBContainer &paContainer) :
+FORTE_ST_DEL_FB::FORTE_ST_DEL_FB(const CStringDictionary::TStringId paInstanceNameId, forte::core::CFBContainer &paContainer) :
     CFunctionBlock(paContainer, &scmFBInterfaceSpec, paInstanceNameId),
     var_conn_QO(var_QO),
     var_conn_STATUS(var_STATUS),
     conn_CNF(this, 0),
     conn_QI(nullptr),
-    conn_ELEM_NAME(nullptr),
+    conn_FB_NAME(nullptr),
     conn_DST(nullptr),
     conn_QO(this, 0, &var_conn_QO),
     conn_STATUS(this, 1, &var_conn_STATUS) {
 };
 
-void FORTE_EC_STOP_ELEM::setInitialValues() {
+void FORTE_ST_DEL_FB::setInitialValues() {
   var_QI = 0_BOOL;
-  var_ELEM_NAME = u""_WSTRING;
+  var_FB_NAME = u""_WSTRING;
   var_DST = u""_WSTRING;
   var_QO = 0_BOOL;
   var_STATUS = u""_WSTRING;
 }
 
-void FORTE_EC_STOP_ELEM::executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) {
+void FORTE_ST_DEL_FB::executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) {
   switch(paEIID) {
     case scmEventREQID:
       var_QO = var_QI;
@@ -75,12 +75,12 @@ void FORTE_EC_STOP_ELEM::executeEvent(TEventID paEIID, CEventChainExecutionThrea
   }
 }
 
-void FORTE_EC_STOP_ELEM::executeRQST(){
+void FORTE_ST_DEL_FB::executeRQST(){
   forte::core::SManagementCMD theCommand;
 
   theCommand.mDestination = CStringDictionary::getInstance().getId(var_DST.getValue());
-  theCommand.mFirstParam.pushBack(CStringDictionary::getInstance().getId(var_ELEM_NAME.getValue()));
-  theCommand.mCMD = EMGMCommandType::Stop;
+  theCommand.mFirstParam.pushBack(CStringDictionary::getInstance().getId(var_FB_NAME.getValue()));
+  theCommand.mCMD = EMGMCommandType::DeleteFBInstance;
 
   EMGMResponse resp = getDevice()->executeMGMCommand(theCommand);
 
@@ -91,10 +91,10 @@ void FORTE_EC_STOP_ELEM::executeRQST(){
   var_STATUS = CIEC_WSTRING(retVal.c_str());
 }
 
-void FORTE_EC_STOP_ELEM::readInputData(TEventID paEIID) {
+void FORTE_ST_DEL_FB::readInputData(TEventID paEIID) {
   switch(paEIID) {
     case scmEventREQID: {
-      readData(1, var_ELEM_NAME, conn_ELEM_NAME);
+      readData(1, var_FB_NAME, conn_FB_NAME);
       readData(2, var_DST, conn_DST);
       readData(0, var_QI, conn_QI);
       break;
@@ -104,7 +104,7 @@ void FORTE_EC_STOP_ELEM::readInputData(TEventID paEIID) {
   }
 }
 
-void FORTE_EC_STOP_ELEM::writeOutputData(TEventID paEIID) {
+void FORTE_ST_DEL_FB::writeOutputData(TEventID paEIID) {
   switch(paEIID) {
     case scmEventCNFID: {
       writeData(1, var_STATUS, conn_STATUS);
@@ -116,16 +116,16 @@ void FORTE_EC_STOP_ELEM::writeOutputData(TEventID paEIID) {
   }
 }
 
-CIEC_ANY *FORTE_EC_STOP_ELEM::getDI(size_t paIndex) {
+CIEC_ANY *FORTE_ST_DEL_FB::getDI(size_t paIndex) {
   switch(paIndex) {
     case 0: return &var_QI;
-    case 1: return &var_ELEM_NAME;
+    case 1: return &var_FB_NAME;
     case 2: return &var_DST;
   }
   return nullptr;
 }
 
-CIEC_ANY *FORTE_EC_STOP_ELEM::getDO(size_t paIndex) {
+CIEC_ANY *FORTE_ST_DEL_FB::getDO(size_t paIndex) {
   switch(paIndex) {
     case 0: return &var_QO;
     case 1: return &var_STATUS;
@@ -133,27 +133,28 @@ CIEC_ANY *FORTE_EC_STOP_ELEM::getDO(size_t paIndex) {
   return nullptr;
 }
 
-CEventConnection *FORTE_EC_STOP_ELEM::getEOConUnchecked(TPortId paIndex) {
+CEventConnection *FORTE_ST_DEL_FB::getEOConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_CNF;
   }
   return nullptr;
 }
 
-CDataConnection **FORTE_EC_STOP_ELEM::getDIConUnchecked(TPortId paIndex) {
+CDataConnection **FORTE_ST_DEL_FB::getDIConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_QI;
-    case 1: return &conn_ELEM_NAME;
+    case 1: return &conn_FB_NAME;
     case 2: return &conn_DST;
   }
   return nullptr;
 }
 
-CDataConnection *FORTE_EC_STOP_ELEM::getDOConUnchecked(TPortId paIndex) {
+CDataConnection *FORTE_ST_DEL_FB::getDOConUnchecked(TPortId paIndex) {
   switch(paIndex) {
     case 0: return &conn_QO;
     case 1: return &conn_STATUS;
   }
   return nullptr;
 }
+
